Add new files to better understand and test the grammar
Added a file example.txt that has all the rules in the grammar. Its aim is to be used to better understand what is and isn't allowed, and when each rule is applied/recognized.

Also added resumen.txt, which explains some of the rules so that they are better/more quickly undestood.